Create AppbarLogo widget and display in all appbars
